inference/pyproject.toml
[project]
name = "inference"
version = "0.1.0"
description = "Two-step therapy response inference service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "google-genai>=1.0.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-frontmatter>=1.1.0",
]
